Reduced warning messages

+++ Solver/Common.h
@@ -61,7 +61,7 @@ namespace Solver
 			size_t result = 0;
 			const size_t prime = 31;
 			int len = strlen(s1);
-			for (size_t i = 0; i < len; ++i) {
+			for (int i = 0; i < len; ++i) {
 				result = s1[i] + (result * prime);
 			}
 			return result;
@@ -82,7 +82,7 @@ namespace Solver
 			const size_t prime = 31;
 			const char* s1 = nd->Positions.get();
 			int len = strlen(s1);
-			for (size_t i = 0; i < len; ++i) {
+			for (int i = 0; i < len; ++i) {
 				result = s1[i] + (result * prime);
 			}
 			return result;
